allow both xml and json from one process
Currently just uses json for "local" styles that resolve to a zotero.org
url and we expect to have locally converted if we're choosing this mode.
Remote styles will be assumed xml.

This is accomplished by having a second citeproc module that adds the CSL_JSON
implementation at the top, and uses that for XML_PARSING in nodejs usage. Also
add separate locale and cslFetcher instances to the engineCache so it can
instantiate the different citeproc engines correctly.

Also treat sigterm the same as sigint for graceful(ish) shutdown.